Fixed some bugs
There was a couple of bad bugs, mostly around registers (we were not
setting them properly...). Ended up doing a little bit of refactoring as
well.

The boot sequence is now looking much better, but is now stuck in an
infinite loop waiting for a screen frame as we haven't implemented any
of the GFX subsystem.