Finally fixed the corrupted graphics in the boot logo
Turns out the boot rom writes tile data at the beginning and relies on
the LCD being of to be able to do so, but I had messed up handling of
that part so the LCD still looped through all the different modes. When
the boot rom tried to access the VRAM during Mode 3, the writes were
ignored causing corruption in the data.